feat: multi-chain contract integration in DAO edit settings
