+++ src/components/settings_and_members/settings.tsx
@@ -1,18 +1,11 @@
 import {
-  Flex, Text, Box, useToast, ToastId,
+  Flex, Text, Box,
 } from '@chakra-ui/react';
-import React, { useEffect, useContext } from 'react';
-import { useContract, useSigner } from 'wagmi';
+import React, { useEffect } from 'react';
 import { Workspace } from 'src/types';
-import useIsCorrectNetworkSelected from 'src/hooks/useIsCorrectNetwork';
-import { CHAIN_INFO } from 'src/constants/chainInfo';
-import { useToastContext } from 'src/context/toastContext';
-import config from '../../constants/config';
-import WorkspaceRegistryABI from '../../contracts/abi/WorkspaceRegistryAbi.json';
+import { useDaoContext } from 'src/context/daoContext';
 import EditForm from './edit_form';
-import { getUrlForIPFSHash, uploadToIPFS } from '../../utils/ipfsUtils';
-import { ApiClientsContext } from '../../../pages/_app';
-import InfoToast from '../ui/infoToast';
+import { getUrlForIPFSHash } from '../../utils/ipfsUtils';
 
 interface Props {
   workspaceData: Workspace;
@@ -33,40 +26,7 @@ function Settings({
     telegramChannel?: string;
   } | null>();
 
-  const [signerStates] = useSigner();
-  const { showErrorToast } = useToastContext();
-
-  const contract = useContract({
-    addressOrName: config.WorkspaceRegistryAddress,
-    contractInterface: WorkspaceRegistryABI,
-    signerOrProvider: signerStates.data,
-  });
-
-  const apiClients = useContext(ApiClientsContext);
-
-  const [hasClicked, setHasClicked] = React.useState(false);
-  const toastRef = React.useRef<ToastId>();
-  const toast = useToast();
-
-  const closeToast = () => {
-    if (toastRef.current) {
-      toast.close(toastRef.current);
-    }
-  };
-
-  const showToast = ({ link } : { link: string }) => {
-    toastRef.current = toast({
-      position: 'top',
-      render: () => (
-        <InfoToast
-          link={link}
-          close={closeToast}
-        />
-      ),
-    });
-  };
-  const isCorrectNetworkSelected = useIsCorrectNetworkSelected(apiClients?.chainId);
-
+  const { editWorkspace, hasClicked } = useDaoContext();
   const handleFormSubmit = async (data: {
     name: string;
     about: string;
@@ -76,65 +36,7 @@ function Settings({
     discordHandle?: string;
     telegramChannel?: string;
   }) => {
-    if (!apiClients) return;
-    try {
-      const { validatorApi } = apiClients;
-
-      if (!isCorrectNetworkSelected && apiClients && apiClients.chainId) {
-        throw Error(`You are on the wrong network. Please switch to ${CHAIN_INFO[apiClients.chainId].name}`);
-      }
-
-      let imageHash = workspaceData.logoIpfsHash;
-      let coverImageHash = workspaceData.coverImageIpfsHash;
-      const socials = [];
-
-      if (data.image) {
-        imageHash = (await uploadToIPFS(data.image)).hash;
-      }
-      if (data.coverImage) {
-        coverImageHash = (await uploadToIPFS(data.coverImage)).hash;
-      }
-
-      if (data.twitterHandle) {
-        socials.push({ name: 'twitter', value: data.twitterHandle });
-      }
-      if (data.discordHandle) {
-        socials.push({ name: 'discord', value: data.discordHandle });
-      }
-      if (data.telegramChannel) {
-        socials.push({ name: 'telegram', value: data.telegramChannel });
-      }
-
-      setHasClicked(true);
-      const {
-        data: { ipfsHash },
-      } = await validatorApi.validateWorkspaceUpdate(coverImageHash ? {
-        title: data.name,
-        about: data.about,
-        logoIpfsHash: imageHash,
-        coverImageIpfsHash: coverImageHash,
-        socials,
-      } : {
-        title: data.name,
-        about: data.about,
-        logoIpfsHash: imageHash,
-        socials,
-      });
-
-      const workspaceID = Number(workspaceData?.id);
-
-      const txn = await contract.updateWorkspaceMetadata(workspaceID, ipfsHash);
-      const transactionData = await txn.wait();
-      setHasClicked(false);
-      window.location.reload();
-
-      showToast({ link: `https://etherscan.io/tx/${transactionData.transactionHash}` });
-    } catch (error: any) {
-      setHasClicked(false);
-      // console.log(error);
-      showErrorToast(error.message as string);
-    }
-    // await subgraphClient.waitForBlock(transactionData.blockNumber);
+    await editWorkspace(data, workspaceData);
   };
 
   useEffect(() => {
